requirements.txt
click==8.1.3
mlflow==2.14.2
mlflow_skinny==2.15.1
numpy==2.1.1
pandas==2.2.2
Pillow==10.4.0
rich==13.8.1
scikit_learn==1.5.2
streamlit==1.37.1
typing_extensions==4.12.2
zenml==0.66.0
